# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rare District — Nigerian fashion marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors/apply:
    post:
      operationId: applyAsVendor
      tags: [Vendors]
      summary: Apply to become a vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorApplication"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Already a vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /vendors/me:
    get:
      operationId: getMyVendorProfile
      tags: [Vendors]
      summary: Get own vendor profile
      responses:
        "200":
          description: Vendor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
        "404":
          description: Not a vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateMyVendorProfile
      tags: [Vendors]
      summary: Update own vendor profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
...
  /vendors/dashboard/recent-orders:
    get:
      operationId: getVendorRecentOrders
      tags: [Vendors]
      summary: Get vendor's recent orders
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent orders for vendor
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [Vendors]
      summary: Get a vendor's public profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ── PRODUCTS ──────────────────────────────────────────────────────────────
...
  /wardrobe:
    get:
      operationId: getWardrobe
      tags: [Wardrobe]
      summary: Get current user's wardrobe items
      responses:
        "200":
          description: Wardrobe items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WardrobeItem"

    post:
      operationId: addToWardrobe
      tags: [Wardrobe]
      summary: Add a product to wardrobe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WardrobeInput"
      responses:
        "201":
          description: Added to wardrobe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WardrobeItem"
  /wardrobe/{productId}:
    delete:
      operationId: removeFromWardrobe
      tags: [Wardrobe]
      summary: Remove a product from wardrobe
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ── ORDERS ────────────────────────────────────────────────────────────────
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an uploaded object entity
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

# ── SCHEMAS ──────────────────────────────────────────────────────────────────
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    User:
      type: object
      required: [id, email, role, referralCode, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [shopper, vendor, admin]
        referralCode:
          type: string
        referredBy:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Vendor:
      type: object
      required: [id, userId, brandName, status, payoutBalance, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        brandName:
          type: string
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
        accountNumber:
          type: ["string", "null"]
        accountName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        commissionRateOverride:
          type: ["number", "null"]
        payoutBalance:
          type: number
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
    VendorApplication:
      type: object
      required: [brandName]
      properties:
        brandName:
          type: string
          minLength: 2
        description:
          type: string
        logoUrl:
          type: string
        website:
          type: string
        bankName:
          type: string
        accountNumber:
          type: string
        accountName:
          type: string
    VendorUpdate:
      type: object
      properties:
        brandName:
          type: string
        description:
          type: string
        logoUrl:
          type: string
        website:
          type: string
        bankName:
          type: string
        accountNumber:
          type: string
        accountName:
          type: string
...
    Product:
      type: object
      required: [id, vendorId, name, price, currency, stock, isActive, createdAt]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        currency:
          type: string
          default: NGN
        category:
          type: ["string", "null"]
        sizes:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        wardrobeCount:
          type: integer
        averageRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        vendor:
          $ref: "#/components/schemas/Vendor"
...
    WardrobeItem:
      type: object
      required: [id, userId, productId, addedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        productId:
          type: integer
        selectedSize:
          type: ["string", "null"]
        quantity:
          type: integer
        addedAt:
          type: string
          format: date-time
        product:
          $ref: "#/components/schemas/Product"
    WardrobeInput:
      type: object
      required: [productId]
      properties:
        productId:
          type: integer
        selectedSize:
          type: string
        quantity:
          type: integer
          default: 1

    # ── Orders ────────────────────────────────────────────────────────────────
    Order:
      type: object
      required: [id, userId, status, totalAmount, currency, shippingAddress, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [pending, paid, processing, shipped, delivered, cancelled]
        totalAmount:
          type: number
        discountAmount:
          type: number
        currency:
          type: string
          default: NGN
        shippingAddress:
          type: string
        shippingCity:
          type: string
        shippingState:
          type: string
        shippingPhone:
          type: string
        couponCode:
          type: ["string", "null"]
        paymentProcessor:
          type: ["string", "null"]
          enum: [paystack, flutterwave, null]
        paymentReference:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderItem:
      type: object
      required: [id, orderId, productId, vendorId, quantity, unitPrice, commissionAmount, vendorAmount]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        vendorId:
          type: integer
        quantity:
          type: integer
        selectedSize:
          type: ["string", "null"]
        unitPrice:
          type: number
        commissionRate:
          type: number
        commissionAmount:
          type: number
        vendorAmount:
          type: number
        product:
          $ref: "#/components/schemas/Product"
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string

        objectPath:
          type: string